online update
git-svn-id: svn://zwischenwelt.org:3690/zw@687 6ae2b031-6f08-0410-bc4a-c5f2e5668f8d

+++ lib.php
@@ -636,7 +636,7 @@ function UpdateSession ($sid)
 	$fastsession = defined("kFastSession");
 	
 	//todo: uncommented due to multi hunting
-	if (!$fastsession)
+	if (!$fastsession && (time()%30) == 0)
 		sql("DELETE FROM `session` WHERE `lastuse` < ".(time()-kSessionTimeout));
 
 	$gSessionObj = sqlgetobject("SELECT * FROM `session` WHERE `sid` = '".addslashes($sid)."' AND `lastuse` > ".(time()-kSessionTimeout));
